identify -> gen_identify
